Added Intersection functionality
Now we can intersect two roads. To do it, press 'i' and the program will
go on Intersect Tool. Now, click on a road, it will highlight red, then
press another road. The start point of the second road will be matched
to the end point of the other road. An intersection object is created
from this and it is added to the RoadManager's m_intersections array
list.