Fixed an issue of road direction
One way signs work for now, but stop signs don't.
If a one way is set in the opposite direction that the car is going,
then it goes back, but this was having issues before. I fixed this by
defaulting a road that gets its connections overwritten as IN types, so
when there's no OUT types or IN_OUT types in the road's connections,
then this is reset to default road behavior.